pyproject.toml
[project]
name = "weather-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncio>=3.4.3",
    "fastapi>=0.115.13",
    "mlflow==3.1.0",
    "pandas>=2.3.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "uvicorn>=0.34.3",
    "torch==2.2.2",
    "scikit-learn==1.6.1",
    "numpy<2",
    "boto3>=1.38.38",
    "python-dotenv>=1.1.0",
]
